Work around Libero symlink issue on Windows
Windows doesn't have symlink support, so you have to link in the symlink
targets.
This means that for the symlinked files, the `include/mbus_def.v` picks
up the version
with power gating turned on. Since this file is in the no_pwr_gating
directory, checking
and re-disabling it is a fine workaround.