Transmitter should forward the ack/nak control bit
During a successful transmission, control bit 1 is driven by the RX node.
This means the TX node should forward DIN to DOUT if there wasn't a
transmission error.

This fix won't affect the correctness of the sending or receiving node,
since the sender sees the ack/nak bit correctly (it just doesn't forward
it currently) and the receiver ignores it. It will, however, confuse a
node that is snooping the transmission, depending where in the ring the
snooping bus is placed.